requirements.txt
# python==3.6.13
fastapi
uvicorn
nltk
pandas
python-dotenv
motor
spacy
wordninja
contractions
streamlit
pyyaml
boto3
cutelog
python-multipart
# 
# AI/ML
# numpy
# tensorflow
opencv-python 
tensorflow==2.4.4
keras==2.4.3 
numpy==1.19.3 
pillow==7.0.0 
scipy==1.4.1 
h5py==2.10.0 
matplotlib==3.3.2 
keras-resnet==0.2.0
imageai==2.1.6
video2images
# pip install nest-asyncio